requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Admin==1.6.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
alembic==1.13.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.1.3
